+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- ISellerを最適化
+- 販売者のIMerchantReturnPolicyのstrict definition
+
 ### Deprecated
 
 ### Removed